tour , place, plandate 수정 완료

+++ services/tour.service.js
@@ -5,7 +5,7 @@ class TourService {
 
   // 여행 계획 작성
 
-  createTour = async ({ user_id, title, start_date, end_date }) => {
+  createTour = async ({ user_id, title, start_date, end_date, tour_site_id }) => {
     if (!user_id) throw { code: 401, message: '유저ID가 존재하지않습니다.' };
 
     if (!title) throw { code: 401, message: 'title을 입력해주세요.' };
@@ -19,6 +19,7 @@ class TourService {
       title,
       start_date,
       end_date,
+      tour_site_id,
     });
 
     if (!createTourData) throw { code: 401, message: '여행계획 등록이 실패하였습니다.' };
@@ -26,25 +27,28 @@ class TourService {
     return { createTourData, code: 200, message: '여행계획 작성이 완료되었습니다.' };
   };
   //여행계획 조회
-  getTour = async ({ tour_id }) => {
-    // 매개변수 객체로 수정
+  getTour = async ({ tour_site_id, tour_id }) => {
     if (!tour_id) {
       throw { code: 400, message: '실패하였습니다.' };
     }
 
-    const findTour = await this.tourRepository.getTour(tour_id);
+    if (!tour_site_id) {
+      throw { code: 400, message: '실패하였습니다.' };
+    }
+    const findTour = await this.tourRepository.getTour(tour_site_id, tour_id);
 
     return {
-      findTour,
+      data: findTour,
       code: 200,
       message: '여행계획 조회가 완료되었습니다.',
     };
   };
   //여행계획 수정
-  putTour = async ({ tour_id, title, start_date, end_date }) => {
+  putTour = async ({ user_id, tour_id, title, start_date, end_date }) => {
     if (!tour_id) throw { code: 400, message: 'tour_id가 필요합니다.' };
 
     const updatedTour = await this.tourRepository.updateTour({
+      user_id,
       tour_id,
       title,
       start_date,
@@ -56,10 +60,10 @@ class TourService {
     return { code: 200, message: '여행 계획이 성공적으로 수정되었습니다.' };
   };
   //여행계획 삭제
-  deleteTour = async ({ tour_id }) => {
+  deleteTour = async ({ user_id, tour_id }) => {
     if (!tour_id) throw { code: 400, message: 'tour_id가 필요합니다.' };
 
-    const deletedTour = await this.tourRepository.deleteTour({ tour_id });
+    const deletedTour = await this.tourRepository.deleteTour({ user_id, tour_id });
 
     if (!deletedTour) throw { code: 400, message: '여행 계획 삭제에 실패했습니다.' };
